requirements.txt
asgiref==3.3.4
Django==3.2.9
django-cors-headers==3.7.0
django-environ==0.4.5
django-ipware==3.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.3
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
requests~=2.26.0